remove pydantic from fs_pred to the pydantic_schema file and update the code
