+++ pkg/fs_algo/fs_algo/fs_pred_algo.py
@@ -14,11 +14,9 @@ import pandera as pa
 from pandera import Column, DataFrameSchema, Index, Check
 from pandera.typing import Series
 from datetime import datetime
-import re
-from typing import Any, Dict, Tuple, Optional
-from pydantic import BaseModel, field_validator, model_validator, ValidationError
 import importlib.util
 import sys
+from fs_algo.pydantic_schemas import ModelMetadata
 
 # TODO create a function that's flexible/converts user formatted checks (a la fs_prep)
 
@@ -55,62 +53,6 @@ if __name__ == "__main__":
         pred_cfg = yaml.safe_load(file)
     
     mapie_alpha = pred_cfg.get('MAPIE_alpha', None)
-
-    # %% Pydantic model pipeline validation 
-    class ModelMetadata(BaseModel):
-        pipeline: Any
-        mapie: Any
-        X_train_shape: Tuple[int, int]
-        Uncertainty: Optional[Dict[str, Any]] = None
-
-        # 1. Check that X_train_shape is a tuple of 2 ints
-        @field_validator("X_train_shape")
-        def validate_shape(cls, v):
-            if not (isinstance(v, tuple) and len(v) == 2 and all(isinstance(i, int) for i in v)):
-                raise ValueError("X_train_shape must be a tuple of two integers")
-            return v
-
-        # 2. Custom model-level validator to handle Uncertainty logic
-        @model_validator(mode="after")
-        def validate_uncertainty(cls, values):
-            unc = values.Uncertainty
-            if unc is None:
-                return values  # optional
-
-            # forestci block
-            if "forestci" in unc:
-                forestci = unc["forestci"]
-                pattern = r"^ci_(\d{1,2}|[1-9][0-9])$"  # zz ∈ [1–99]
-                matched = [k for k in forestci if re.match(pattern, k)]
-                if not matched:
-                    raise ValueError("forestci must contain at least one 'ci_zz' with zz between 1 and 99")
-                for k in matched:
-                    ci = forestci[k]
-                    for bound in ["upper_bound", "lower_bound"]:
-                        arr = ci.get(bound)
-                        if not (isinstance(arr, np.ndarray) and np.issubdtype(arr.dtype, np.floating)):
-                            raise ValueError(f"forestci -> {k} -> {bound} must be a NumPy array of floats")
-
-            # bagging_confidence_interval block
-            if "bagging_confidence_interval" in unc:
-                for key in ["bagging_std_pred", "bagging_mean_pred"]:
-                    arr = unc.get(key)
-                    if not (isinstance(arr, np.ndarray) and np.issubdtype(arr.dtype, np.floating)):
-                        raise ValueError(f"{key} must be a NumPy array of floats")
-
-                confs = unc.get("bagging_confidence_intervals", {})
-                pattern = r"^confidence_level_(\d{1,2}|[1-9][0-9])$"
-                matched = [k for k in confs if re.match(pattern, k)]
-                if not matched:
-                    raise ValueError("bagging_confidence_intervals must contain keys like 'confidence_level_zz'")
-
-                for k in matched:
-                    for bound in ["upper_bound", "lower_bound"]:
-                        arr = confs[k].get(bound)
-                        if not (isinstance(arr, np.ndarray) and np.issubdtype(arr.dtype, np.floating)):
-                            raise ValueError(f"bagging_confidence_intervals -> {k} -> {bound} must be a NumPy array of floats")
-
-            return values
 
     #%%  READ CONTENTS FROM THE ATTRIBUTE CONFIG
     path_attr_config = fsate.build_cfig_path(path_pred_config,pred_cfg.get('name_attr_config',None))
@@ -210,12 +152,7 @@ if __name__ == "__main__":
                 pipeline_data = joblib.load(path_algo)
                 
                 # Validate joblib file
-                try:
-                    validated = ModelMetadata(**pipeline_data)
-                    print("Validation successful ✅")
-                except ValidationError as e:
-                    print("Validation failed ❌")
-                    print(e)
+                validated = ModelMetadata(**pipeline_data)
                 
                 pipe = pipeline_data['pipeline']
                 X_train_shape = pipeline_data['X_train_shape']  # Retrieve X_train.shape
